Add cli CRUD commands
