+++ pcrmc/view/edit.py
@@ -1,5 +1,5 @@
-"""This module provides the create command"""
-# pcrmc/view/create.py
+"""This module provides the edit command"""
+# pcrmc/view/edit.py
 
 from typing import List
 import typer
@@ -9,12 +9,12 @@ app = typer.Typer()
 
 @app.command()
 def contact(identifier: List[str]):
-    print(f"Creating contact: {identifier}")
+    print(f"Editing contact: {identifier}")
 
 
 @app.command()
 def meeting(identifier: List[str]):
-    print(f"Creating meeting: {identifier}")
+    print(f"Editing meeting: {identifier}")
 
 
 if __name__ == "__main__":
