+++ pcrmc/view/show.py
@@ -1,5 +1,5 @@
-"""This module provides the create command"""
-# pcrmc/view/create.py
+"""This module provides the show command"""
+# pcrmc/view/show.py
 
 from typing import List
 import typer
@@ -9,12 +9,12 @@ app = typer.Typer()
 
 @app.command()
 def contact(identifier: List[str]):
-    print(f"Creating contact: {identifier}")
+    print(f"Showing contact: {identifier}")
 
 
 @app.command()
 def meeting(identifier: List[str]):
-    print(f"Creating meeting: {identifier}")
+    print(f"Showing meeting: {identifier}")
 
 
 if __name__ == "__main__":
